move status polling to commong
