+++ src/app/common/duck/actions.ts
@@ -30,3 +30,12 @@ export const startDataListPolling = (params?: any) => ({
 export const stopDataListPolling = () => ({
   type: 'DATA_LIST_POLL_STOP',
 });
+
+export const startStatusPolling = (params?: any) => ({
+  type: 'STATUS_POLL_START',
+  params,
+});
+
+export const stopStatusPolling = () => ({
+  type: 'STATUS_POLL_STOP',
+});
