+++ src/app/plan/duck/actions.ts
@@ -1,8 +1,8 @@
 import { createActions } from 'reduxsauce';
 
 const { Creators, Types } = createActions({
-  startStatusPolling: ['params'],
-  stopStatusPolling: [],
+  // startStatusPolling: ['params'],
+  // stopStatusPolling: [],
   startPVPolling: ['params'],
   stopPVPolling: [],
   updatePlans: ['updatedPlans'],
