+++ src/app/plan/duck/sagas.ts
@@ -2,33 +2,33 @@ import { race, call, delay, put, take } from 'redux-saga/effects';
 
 import { Creators } from './actions';
 
-function* checkStatus(action) {
-  const params = { ...action.params };
-  while (true) {
-    const plansRes = yield call(params.asyncFetch);
-    const pollingStatus = params.callback(plansRes);
+// function* checkStatus(action) {
+//   const params = { ...action.params };
+//   while (true) {
+//     const plansRes = yield call(params.asyncFetch);
+//     const pollingStatus = params.callback(plansRes);
 
-    switch (pollingStatus) {
-      case 'SUCCESS':
-        yield put({ type: 'STOP_STATUS_POLLING' });
-        break;
-      case 'FAILURE':
-        Creators.stopStatusPolling();
+//     switch (pollingStatus) {
+//       case 'SUCCESS':
+//         yield put({ type: 'STOP_STATUS_POLLING' });
+//         break;
+//       case 'FAILURE':
+//         Creators.stopStatusPolling();
 
-        yield put({ type: 'STOP_STATUS_POLLING' });
-        break;
-      default:
-        break;
-    }
-    yield delay(params.delay);
-  }
-}
-function* watchStatusPolling() {
-  while (true) {
-    const data = yield take(Creators.startStatusPolling().type);
-    yield race([call(checkStatus, data), take('STOP_STATUS_POLLING')]);
-  }
-}
+//         yield put({ type: 'STOP_STATUS_POLLING' });
+//         break;
+//       default:
+//         break;
+//     }
+//     yield delay(params.delay);
+//   }
+// }
+// function* watchStatusPolling() {
+//   while (true) {
+//     const data = yield take(Creators.startStatusPolling().type);
+//     yield race([call(checkStatus, data), take('STOP_STATUS_POLLING')]);
+//   }
+// }
 
 function* checkPVs(action) {
   const params = { ...action.params };
@@ -66,6 +66,6 @@ function* watchPVPolling() {
 }
 
 export default {
-  watchStatusPolling,
+  // watchStatusPolling,
   watchPVPolling,
 };
